+++ documentation/docs/changelog.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ---
 
+## Storefront Catalog Service
+
+### 0.1.1 - 2026-03-08
+[PR](https://github.com/smartguy-coder/PrototypeHighlyLoadedDistributedService/pull/2)
+
+#### Added
+- Created storefront service backend
+- Custom User model with email or phone authentication
+- EmailOrPhoneBackend for flexible login
+
+---
+
 ## Documentation
 
 ### 0.1.0 - 2026-02-10